SIGALRM: dump buffered data
Trailing data is not dumped until the read buffer is full. If the
subprocess writes infrequently, the buffered data may not be written
for some time or when the hexlog process exits.

Add an environment variable (HEXLOG_TIMEOUT) to periodically write out
any buffered data. 0 (the default) disables the timer.